+++ visualization/app/codeCharta/renderer/sunburst/sunburst.facade.ts
@@ -1,4 +1,5 @@
 export { SunburstComponent } from "./components/sunburst/sunburst.component"
 export type { SunburstColoring } from "./util/sunburstColor"
+export { VISIBLE_RING_COUNT } from "./util/sunburstOption.builder"
 export type { SunburstMetrics, SunburstNode } from "./util/sunburstTree"
-export { buildSunburstTree, findClosestFolder, parentPath } from "./util/sunburstTree"
+export { buildSunburstTree, findClosestFolder, findClosestNode, findFolder, parentPath } from "./util/sunburstTree"